requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
matplotlib
plotly
pytest
xgboost
lightgbm
prophet
PyYAML
